+++ EscapeApp/app/src/androidTest/java/com/context/kroket/escapeapp/Game_A_ActivityTest.java
@@ -8,7 +8,8 @@ import android.support.test.espresso.intent.rule.IntentsTestRule;
 import android.support.test.filters.LargeTest;
 import android.support.test.runner.AndroidJUnit4;
 
-import org.junit.Before;
+import com.context.kroket.escapeapp.mainscreens.MainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
