+++ EscapeApp/app/src/main/AndroidManifest.xml
@@ -10,8 +10,8 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
-        android:name=".App">
-        <activity android:name=".MainActivity">
+        android:name=".application.App">
+        <activity android:name=".mainscreens.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,19 +20,19 @@
         </activity>
 
         <activity
-            android:name=".Game_A_Activity"
+            android:name=".minigames.Game_A_Activity"
             android:label="@string/title_activity_game__a_"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".Game_AA_Activity" />
+        <activity android:name=".minigames.Game_AA_Activity" />
 
         <service
-            android:name=".ConnectionService"
+            android:name=".network.ConnectionService"
             android:enabled="true"
             android:exported="true" />
 
-        <activity android:name=".Game_B_Activity" />
-        <activity android:name=".WaitingActivity" />
-        <activity android:name=".Game_C_Activity"></activity>
+        <activity android:name=".minigames.Game_B_Activity" />
+        <activity android:name=".mainscreens.WaitingActivity" />
+        <activity android:name=".minigames.Game_C_Activity"></activity>
     </application>
 
 </manifest>
