--- EscapeApp/app/src/main/java/com/context/kroket/escapeapp/WaitingActivity.java
@@ -1,34 +1,0 @@
-package com.context.kroket.escapeapp;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-/**
- * Responsible for the screen that is shown when no minigame is active.
- */
-public class WaitingActivity extends AppCompatActivity {
-
-    /**
-     * Initializes the layout.
-     *
-     * @param savedInstanceState If the activity is being re-initialized after
-     *     previously being shut down then this Bundle contains the data it most
-     *     recently supplied.
-     */
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_waiting);
-    }
-
-    /**
-     * Start the activity. Change current activity to this.
-     */
-    @Override
-    protected void onStart() {
-        super.onStart();
-
-        //Change the current activity
-        ((App)this.getApplicationContext()).setCurrentActivity(this);
-    }
-}
