+++ EscapeApp/app/src/main/res/layout/activity_game__c_.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.context.kroket.escapeapp.Game_C_Activity">
+    tools:context=".minigames.Game_C_Activity">
 
     <TextView android:id="@+id/gameC"
         android:layout_width="wrap_content"
