+++ EscapeApp/app/src/main/res/layout/content_game__a_.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.context.kroket.escapeapp.Game_A_Activity"
+    tools:context=".minigames.Game_A_Activity"
     tools:showIn="@layout/activity_game_a">
 
 </RelativeLayout>
